start the integration test "infrastructure"
There's no test yet, only the system to bootstrap the docker images that are
required to execute tests.

The general idea is to have multiple configuration files in the test/assets
directory that are used to simulate different configurations. These configurations
are executed from a docker image running a local copy of xivo-dird.

The Dockerfiles are in contrib/docker/base_test_image and tests

The base_test_image is used to speed up the image generation of the dird-test image

Before testing xivo-dird you must build the base_test_image

make test-setup

Then you can rebuild you test image when you change your local copy of xivo-dird

make test-image

There's an example config in tests/assets/test_csv_backend that can be launched with

make test-csv

This last command is meant as documentation only. The test runner should be able
to launche the docker image with the appropriate configuration file when running.