ldap: change the ldap config behavior
if ldap_custom_filter AND searched_columns are set the resulting filter
will be the custom filter AND the generated filter from the
searched_columns.

if only the ldap_custom_filter is configured, the filter will be the
custom filter only

if only the searched_columns is configured, the filter will be the
generated filter from the searched_columns.

This change fixes the behavior of xivo-dird to have the same results as
the old directory system with changing the configuration.