phone: add fallback fields and multiple results support
Fallback fields are used to provide the same behaviour as the current
search in LDAP filter from phones in XiVO. For a given lookup result,
the phone will use the first key has the name; if the key is not present
(or has a false value), it will try the second key, etc.

Multiple results are used to provide the same behaviour as the current
search in phonebook from phones in XiVO. For a given lookup result, you
can have multiple (e.g. 4) "display results" in the phone directory.

There is still some work to be done: the variables names used in the
module and the config feels poor, and the validity check on the config
could be less cumbersome.