int tests: add no acl authentication coverage
Authentication coverage doesn't need to include all phone plugins,
because they all use the same phone_view plugin (with the same ACL).

The verification of the required_acl has been removed. Since it's not
easy to test, it's more a question of configuration than a
implementation. So, only the "no acl" are verify and not the "right
acl".