fix stuck thread
Scenario: when SSL certificate is missing, the bus thread keeps running...

I tried adapting the test_https.py integration tests, but there's
another bug preventing the test to pass: when xivo-dird is still
connecting to the bus, the exception is raised, the bus thread is
asked to stop, but it does not, because it is still trying to
establish a connection with rabbitmq. This is a problem in the
ConsumerMixin of kombu.