remove divergent id contact fields where id != contact_uuid
reason: when mass-importing contact, it is possible to specify an id
key for contacts, which will be different from the generated
contact_uuid. The contact will then have a "wrong" id contact
field. Until now, it did not have much consequence since the correct
id contact field was duplicated for every contact field and the
"wrong" id field would not be used. But now there is only one correct
id contact field, and the wrong one may override it much more easily