refactor: stop using tenant.name internally
only the phonebook interface uses the tenant name. it then fetches tenants from wazo-auth to find the uuid
of the tenant.

the tenant.name field will be removed after a point in time when we know that all database have been upgraded
to have the same uuid in dird and auth. This will most likely be after the migration to Debian Buster