tenant: add a plugin to change the uuid of a tenant
This plugin is used to change the uuid of tenants in dird to match tenants in wazo-auth.

The plugin is meant to be used by wazo-upgrade in a post upgrade script after the migration, this plugin
should be disabled.

After the next "forced" upgrade, probably for Debian Buster, this plugin could be deleted