phone lookup: use the tenant of the user specified in the URL
phones not authenticate to do lookup queries, dird_phoned is used as a proxy and uses
a token which has the master tenant.

Using that tenant would result in a 404 because the profile cannot be found
from this UUID. Using recurse is not an options since many profiles can
have the same name in different tenants.

the user uuid in the URL is used to query wazo-auth and find the tenant
of the user owning the phone