profile: implement PUT /profiles
adds new db fields where there are relationships to add constraints on
the tenant uuid of related object.

for example

the profile table now has a display_uuid and a display_tenant_uuid field which
if a foreign key on the display table.

by adding the constraint that profile.tenant_uuid equals profile.display_tenant_uuid
we can ensure that it is impossible to make invalid associations