personal: use the uuid from metadata
if xivo_user_uuid field is deprecated and should not be used, I chose the
metadata['uuid'] field to allow any wazo-auth user to have contacts, not only
pbx users which in most cases is the same.

If a token has no metadata['uuid'], which is not possible at the moment, a 401
error is returned instead of listing all personal contacts.